% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest.project.filename.R
\name{escape.regex}
\alias{escape.regex}
\title{Escapes regular expression metacharacters in a string with by prepending `\\Q` and appending `\\E`}
\usage{
escape.regex(x)
}
\arguments{
\item{x}{String with possible regular expression metacharacters}
}
\value{
String `\\Qx\\E`
}
\description{
See the documentation on `base::regex`
}
\examples{
grepl(escape.regex('[abc]'), '[abc]')
grepl(escape.regex('.*'), '.*')
}
