\name{is.duplicated}
\alias{is.duplicated}
\title{
Check for duplicated values
}
\description{
Check for duplicated values.  Logical is returned for all replicates.
}
\usage{
is.duplicated(x)
}
\arguments{
  \item{x}{A vector}
}
\details{
}
\value{
A logical for all replicates, including the first value.
}
\author{
Daryl Waggott
}

\examples{

}
\keyword{ duplicate }
