% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest.project.filename.R
\name{latest.project.filename}
\alias{latest.project.filename}
\title{Inverse of `generate.filename`}
\usage{
latest.project.filename(
  project.stem,
  file.core,
  extension,
  folder.path = ".",
  file.date = NULL
)
}
\arguments{
\item{project.stem}{Name of the project}

\item{file.core}{Main part of the filename}

\item{extension}{What type of file is this}

\item{file.date}{A specific date of the file be located. Leave NULL to find latest file.}
}
\description{
Inverse of `generate.filename`
}
\examples{
file.create(generate.filename('TestProject', 'myfile', 'txt', file.date = '2022-12-12'))
file.create(generate.filename('TestProject', 'myfile', 'txt'))
file.create(generate.filename('TestProject', 'myfile', 'txt', file.date = FALSE))
latest.project.filename('TestProject', 'myfile', 'txt')
}
