\name{mixed.index}
\alias{mixed.index}
\title{Mixed index}
\description{Use forward and reverse indices.}
\usage{
mixed.index(x, index)
}
\arguments{
    \item{x}{A vector}
    \item{index}{A index where values prefixed by a "-" are counted from reverse.}
}
\details{}
\value{}
\author{Daryl Waggott}
\examples{}
\keyword{index}
