\name{read.tabdelimited}
\alias{read.tabdelimited}
\title{Read a Tab-Delimited File}
\description{The function reads a tab-delimited file that has a header and row-names}
\usage{
read.tabdelimited(file);
}
\arguments{
    \item{file}{Name of the file to be read}
}
\details{This is a very simple wrapper function, and should *not* be extended!}
\value{Returns the results of a read.table() call}
\author{Paul Boutros}
\examples{# Add examples!}
\keyword{utilities}
