% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.filename.R
\name{generate.filename}
\alias{generate.filename}
\title{Generate a filename in the lab standard}
\usage{
generate.filename(project.stem, file.core, extension, file.date = Sys.Date())
}
\arguments{
\item{project.stem}{Name of the project}

\item{file.core}{Main part of the filename}

\item{extension}{File extension}

\item{file.date}{Optional date of filename. (defaults to today; FALSE or NULL turns off the date-stamp)}
}
\value{
Boutros lab standard filename
}
\description{
This function creates a filename according to the date_project_core.extension lab standard. Matches the regular expression `r valid.filename.regex`
}
\examples{
# generate a filename to screen
generate.filename('NSCLC', 'StatisticalAnalysis', 'txt');

save.session.profile( generate.filename('Prostate', 'StatisticalAnalysisSessionProfile', 'txt') );

generate.filename('NSCLC', 'StatisticalAnalysis', 'txt', FALSE);
}
